requirements.txt
fastapi==0.79.0
argon2-cffi==21.3.0
itsdangerous==2.1.2
sentry-sdk==1.9.5
beanie==1.11.7
slowapi==0.1.5
msgspec==0.8.0
aiokafka==0.7.2
python-dotenv==0.20.0
email-validator==1.2.1
uvicorn[standard]==0.18.2